x86/traps: Move control protection handler to separate file
commit 2da5b91fe4092aab9d92138c78b68e9856b078d0 upstream.

Today the control protection handler is defined in traps.c and used only
for the kernel IBT feature. To reduce ifdeffery, move it to it's own file.
In future patches, functionality will be added to make this handler also
handle user shadow stack faults. So name the file cet.c.

No functional change.

Intel-SIG: commit 2da5b91fe409 x86/traps: Move control protection
handler to separate file.

Acked-by: Mike Rapoport (IBM) <rppt@kernel.org>
Tested-by: Pengfei Xu <pengfei.xu@intel.com>
Tested-by: John Allen <john.allen@amd.com>
Tested-by: Kees Cook <keescook@chromium.org>
Link: https://lore.kernel.org/all/20230613001108.3040476-8-rick.p.edgecombe%40intel.com

Conflicts:
    arch/x86/kernel/Makefile
    arch/x86/kernel/cet.c
    arch/x86/kernel/traps.c
[ Xingjiang Lu: resolve context conflicts ]
